Move `JsonApiRouterBuilder` to separate module
It's starting to get a bit crowded in `iron/mod.rs`. Since
`JsonApiRouterBuilder` is the biggest chunk in that file, and since it
has nothing to do with the other things in `mod.rs` let's move that
file.
